fix: reuse participant interface
